requirements.txt
ace_tools==0.0
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
audioread==3.0.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
comtypes==1.4.10
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
distro==1.9.0
executing==2.2.0
feedparser==6.0.11
fonttools==4.55.8
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
jiter==0.8.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.44.0
matplotlib==3.10.0
matplotlib-inline==0.1.7
moviepy==2.1.2
msgpack==1.1.0
nest-asyncio==1.6.0
numba==0.61.0
numpy==2.1.3
openai==1.61.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
praw==7.8.1
prawcore==2.4.0
proglog==0.1.10
prompt_toolkit==3.0.50
psutil==6.1.1
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.1
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.98
pytz==2025.1
pywin32==308
pyzmq==26.2.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
setuptools==75.8.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
stack-data==0.6.3
threadpoolctl==3.5.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
update-checker==0.18.0
urllib3==2.3.0
wcwidth==0.2.13
websocket-client==1.8.0
wheel==0.45.1
